modified func7

+++ func7.cpp
@@ -2,10 +2,10 @@
 #include "func7.h"
 using namespace std;
 
-void factorial(int number) {
+void factorial(double number) {
 
-    int unsigned fact = 1;
-    int unsigned i = 1;
+    double fact = 1;
+    double i = 1;
     for (; i <= number; i++) {
         fact = fact * i;
     }
